pyproject.toml
[project]
name = "elixir"
version = "0.1.0"
description = "Multi-user personal finance manager"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.136.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "pydantic-settings>=2.0.0",
    "pyjwt>=2.8.0",
    "bcrypt>=4.1.0",
    "cryptography>=42.0.0",
    "temporalio>=1.6.0",
    "httpx>=0.27.0",
    "python-multipart>=0.0.9",
    "google-adk>=1.31.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/elixir"]

[dependency-groups]
dev = [
    "mypy>=1.20.2",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.15.12",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
